# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Me": {
        "properties": {
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "emailVerified": {
            "title": "Emailverified",
            "type": "boolean"
          },
          "fullName": {
            "title": "Fullname",
            "type": "string"
          },
          "historyEnabled": {
            "title": "Historyenabled",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "marketingOptIn": {
            "title": "Marketingoptin",
            "type": "boolean"
          },
          "permissions": {
            "items": {
              "type": "string"
            },
            "title": "Permissions",
            "type": "array"
          },
          "profile": {
            "$ref": "#/components/schemas/Profile-Output"
          },
          "roles": {
            "items": {
              "type": "string"
            },
            "title": "Roles",
            "type": "array"
          }
        },
        "required": [
          "id",
          "email",
          "fullName",
          "emailVerified",
          "historyEnabled",
          "marketingOptIn",
          "roles",
          "permissions",
          "profile",
          "createdAt"
        ],
        "title": "Me",
        "type": "object"
      },
      "MePatch": {
        "description": "Only the fields sent are changed; within `profile`, too.",
        "properties": {
          "fullName": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "pattern": "^[^\\x00-\\x1f\\x7f]+$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fullname"
          },
          "historyEnabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Historyenabled"
          },
          "marketingOptIn": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Marketingoptin"
          },
          "profile": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Profile-Input"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "MePatch",
        "type": "object"
      },
...
      "Page_ViewOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ViewOut"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "pageSize": {
            "title": "Pagesize",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "pageSize"
        ],
        "title": "Page[ViewOut]",
        "type": "object"
      },
...
      "PasswordIn": {
        "properties": {
          "password": {
            "maxLength": 1000,
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "password"
        ],
        "title": "PasswordIn",
        "type": "object"
      },
...
      "ViewIn": {
        "properties": {
          "propertyId": {
            "title": "Propertyid",
            "type": "string"
          }
        },
        "required": [
          "propertyId"
        ],
        "title": "ViewIn",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Ireland's Property Price Register, made usable. Contains Residential Property Price Register data \u00a9 Property Services Regulatory Authority; it may contain errors.",
    "title": "PPR Map API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/me": {
      "delete": {
        "description": "Close the account now; its data is purged after 30 days (`app.cli purge-deleted`).\nThe last active admin cannot (409): someone must be able to run the site.",
        "operationId": "delete_me_api_v1_me_delete",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Me",
        "tags": [
          "me"
        ]
      },
      "get": {
        "operationId": "get_me_api_v1_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Me"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Me",
        "tags": [
          "me"
        ]
      },
      "patch": {
        "operationId": "patch_me_api_v1_me_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MePatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Me"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Patch Me",
        "tags": [
          "me"
        ]
      }
    },
...
    "/api/v1/me/history/views": {
      "delete": {
        "operationId": "clear_views_api_v1_me_history_views_delete",
        "responses": {
          "204": {
            "description": "Successful Response"
          }
        },
        "summary": "Clear Views",
        "tags": [
          "me"
        ]
      },
      "get": {
        "description": "Property pages visited in the last 12 months, newest first.",
        "operationId": "views_api_v1_me_history_views_get",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "maximum": 100,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "pageSize",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Pagesize",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_ViewOut_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Views",
        "tags": [
          "me"
        ]
      },
      "post": {
        "description": "Record a property page visit, unless history is switched off (then nothing is kept).",
        "operationId": "record_view_api_v1_me_history_views_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ViewIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Record View",
        "tags": [
          "me"
        ]
      }
    },
    "/api/v1/me/history/views/{view_id}": {
      "delete": {
        "operationId": "delete_view_api_v1_me_history_views__view_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "view_id",
            "required": true,
            "schema": {
              "title": "View Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete View",
        "tags": [
          "me"
        ]
      }
    },
    "/api/v1/me/password": {
      "post": {
        "description": "Change the password; every other session is signed out and reset links stop working.",
        "operationId": "change_password_api_v1_me_password_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordChangeIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Change Password",
        "tags": [
          "me"
        ]
      }
    },
...
    "/api/v1/me/wishlist": {
      "get": {
        "operationId": "wishlist_api_v1_me_wishlist_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/WishlistItemOut"
                  },
                  "title": "Response Wishlist Api V1 Me Wishlist Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Wishlist",
        "tags": [
          "me"
        ]
      },
      "post": {
        "description": "Save a property or an area (201). Saving the same one again returns the existing item\n(200), with its note replaced if a new one is given.",
        "operationId": "add_to_wishlist_api_v1_me_wishlist_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WishlistIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WishlistItemOut"
                }
              }
            },
            "description": "Already saved"
          },
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WishlistItemOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add To Wishlist",
        "tags": [
          "me"
        ]
      }
    },
    "/api/v1/me/wishlist/compare": {
      "get": {
        "description": "Side by side: the hover summaries of up to four saved properties, in the order given.",
        "operationId": "compare_api_v1_me_wishlist_compare_get",
        "parameters": [
          {
            "description": "Up to 4 property ids from the wishlist, comma-separated",
            "in": "query",
            "name": "ids",
            "required": true,
            "schema": {
              "description": "Up to 4 property ids from the wishlist, comma-separated",
              "title": "Ids",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PropertySummary"
                  },
                  "title": "Response Compare Api V1 Me Wishlist Compare Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Compare",
        "tags": [
          "me"
        ]
      }
    },
    "/api/v1/me/wishlist/{item_id}": {
      "delete": {
        "operationId": "remove_wishlist_item_api_v1_me_wishlist__item_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "title": "Item Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Remove Wishlist Item",
        "tags": [
          "me"
        ]
      },
      "patch": {
        "operationId": "note_wishlist_item_api_v1_me_wishlist__item_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "title": "Item Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WishlistNote"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WishlistItemOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Note Wishlist Item",
        "tags": [
          "me"
        ]
      }
    },
...
  }
}